Push extraction of languages to `dashboard-app.ts`
Feature story:
A teacher should only see available translations for a glossary in the language settings of the dashboard.

[#169452886]
https://www.pivotaltracker.com/story/show/169452886
